Added default types for Models, updated use of model configs, refactored factory to separate methods into submodules
