+++ mlte/negotiation/model.py
@@ -50,7 +50,10 @@ class SystemDescriptor(BaseModel):
 class NegotiationCardModel(BaseModel):
     """The model implementation for the NegotiationCard artifact."""
 
-    artifact_type: Literal[ArtifactType.NEGOTIATION_CARD]
+    artifact_type: Literal[
+        ArtifactType.NEGOTIATION_CARD
+    ] = ArtifactType.NEGOTIATION_CARD
+
     """Union discriminator."""
 
     system: SystemDescriptor = SystemDescriptor()
