+++ mlte/store/artifact/underlying/rdbs/metadata_spec.py
@@ -1,7 +1,8 @@
 """
 mlte/store/artifact/underlying/rdbs/metadata_spec.py
 
-Definition of the metadata (DB schema) for spec and validated spec in the artifact store.
+Definition of the metadata (DB schema) for spec and validated spec in the artifact store,
+as well as converstions between schema and internal models.
 """
 from __future__ import annotations
