+++ mlte/value/base.py
@@ -62,7 +62,6 @@ class ValueBase(artifact.Value, metaclass=abc.ABCMeta):
         return ArtifactModel(
             header=self.build_artifact_header(),
             body=ValueModel(
-                artifact_type=ArtifactType.VALUE,
                 metadata=self.metadata,
                 value_class=self.get_class_path(),
                 value=OpaqueValueModel(
