+++ mlte/value/types/image.py
@@ -59,7 +59,6 @@ class Image(Value):
         return ArtifactModel(
             header=self.build_artifact_header(),
             body=ValueModel(
-                artifact_type=ArtifactType.VALUE,
                 metadata=self.metadata,
                 value_class=self.get_class_path(),
                 value=ImageValueModel(
