+++ mlte/value/types/real.py
@@ -43,7 +43,6 @@ class Real(Value):
         return ArtifactModel(
             header=self.build_artifact_header(),
             body=ValueModel(
-                artifact_type=ArtifactType.VALUE,
                 metadata=self.metadata,
                 value_class=self.get_class_path(),
                 value=RealValueModel(
